Extract PDF download modal into ReportPDFDownloadModal component
